requirements.txt
pillow
pygame
selenium
websockets
